api/microservices/steam/requirements.txt
grpc-interceptor ~= 0.11.0
flask ~= 1.1
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
pytest ~= 5.4
markupsafe ~= 2.0.1
